backend/requirements.txt
pdfplumber==0.9.0
flask==2.2.5
flask-cors==3.0.10
pytesseract==0.3.10
Pillow==10.0.0
numpy<=1.24.3
torch==2.0.1
transformers==4.30.2
werkzeug==2.2.3